# openapi/fc_openapi.yaml
openapi: 3.0.1
info:
  title: Eclipse XFSC Federated Catalogue
  description: REST API of the XFSC catalogue
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: https://fc-server.xfsc.org
    description: some future test environment
components:
  responses:
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: >-
        Forbidden. The user does not have the permission to execute this
        request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServerError:
      description: >-
        May contain hints how to solve the error or indicate what went wrong at
        the server. Must not outline any information about the internal
        structure of the server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
...
    TrustFrameworkEntry:
      type: object
      properties:
        id:
          type: string
          description: Trust framework identifier
        name:
          type: string
          description: Display name
        serviceUrl:
          type: string
          description: Trust framework service URL
        apiVersion:
          type: string
          description: API version
        timeoutSeconds:
          type: integer
          description: Connection timeout in seconds
        connected:
          type: boolean
          description: Whether the service is currently reachable
        enabled:
          type: boolean
          description: Whether this framework is active
    TrustFrameworkConfigUpdate:
      type: object
      properties:
        serviceUrl:
          type: string
          description: New service URL
        apiVersion:
          type: string
          description: New API version
        timeoutSeconds:
          type: integer
          description: New timeout in seconds
...
    AssetTypeConfig:
      type: object
      properties:
        enabled:
          type: boolean
          description: Whether asset type restriction is active
        allowedTypes:
          type: array
          items:
            type: string
          description: List of allowed asset types (JSON-LD @type URIs)
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /admin/trust-frameworks:
    get:
      tags:
        - TrustFrameworkAdmin
      summary: List all registered trust frameworks
      operationId: getTrustFrameworks
      security:
        - jwt: []
      responses:
        '200':
          description: List of trust frameworks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrustFrameworkEntry'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/trust-frameworks/{id}/enabled:
    put:
      tags:
        - TrustFrameworkAdmin
      summary: Toggle a trust framework's enabled state
      operationId: setTrustFrameworkEnabled
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: enabled
          in: query
          required: true
          schema:
            type: boolean
      responses:
        '200':
          description: Framework enabled state updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/trust-frameworks/{id}:
    put:
      tags:
        - TrustFrameworkAdmin
      summary: Update a trust framework's configuration
      operationId: updateTrustFrameworkConfig
      security:
        - jwt: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrustFrameworkConfigUpdate'
      responses:
        '200':
          description: Framework configuration updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /admin/asset-types:
    get:
      tags:
        - AssetTypeAdmin
      summary: Get asset type restriction configuration
      operationId: getAssetTypeConfig
      security:
        - jwt: []
      responses:
        '200':
          description: Asset type restriction configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetTypeConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - AssetTypeAdmin
      summary: Update asset type restriction configuration
      operationId: updateAssetTypeConfig
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetTypeConfig'
      responses:
        '200':
          description: Configuration updated
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
  /admin/asset-types/existing:
    get:
      tags:
        - AssetTypeAdmin
      summary: Get distinct asset types from stored assets
      operationId: getExistingAssetTypes
      security:
        - jwt: []
      responses:
        '200':
          description: List of distinct asset types
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/ServerError'
...